Put tests in separate files
